+++ router-tests/testenv/testenv.go
@@ -120,8 +120,8 @@ type Config struct {
 	DisableParentBasedSampler          bool
 	TLSConfig                          *core.TlsConfig
 	TraceExporter                      trace.SpanExporter
-	OtelAttributes                     []config.OtelAttribute
-	OtelResourceAttributes             []config.OtelResourceAttribute
+	OtelAttributes                     []config.CustomAttribute
+	OtelResourceAttributes             []config.CustomStaticAttribute
 	MetricReader                       metric.Reader
 	PrometheusRegistry                 *prometheus.Registry
 	ShutdownDelay                      time.Duration
@@ -129,6 +129,11 @@ type Config struct {
 	PropagationConfig                  config.PropagationConfig
 	CacheControlPolicy                 config.CacheControlPolicy
 	LogObservation                     LogObservationConfig
+	ClientHeader                       config.ClientHeader
+	ResponseTraceHeader                config.ResponseTraceHeader
+	Logger                             *zap.Logger
+	AccessLogger                       *zap.Logger
+	AccessLogFields                    []config.CustomAttribute
 }
 
 type SubgraphsConfig struct {
@@ -415,24 +420,27 @@ func createTestEnv(t testing.TB, cfg *Config) (*Environment, error) {
 	}
 
 	var (
-		zapLogger   *zap.Logger
 		logObserver *observer.ObservedLogs
 	)
 
 	if oc := cfg.LogObservation; oc.Enabled {
 		var zCore zapcore.Core
 		zCore, logObserver = observer.New(oc.LogLevel)
-		zapLogger = logging.NewZapLoggerWithCore(zCore, true)
+		cfg.Logger = logging.NewZapLoggerWithCore(zCore, true)
 	} else {
 		ec := zap.NewProductionEncoderConfig()
 		ec.EncodeDuration = zapcore.SecondsDurationEncoder
 		ec.TimeKey = "time"
 
 		syncer := zapcore.AddSync(os.Stderr)
-		zapLogger = logging.NewZapLoggerWithSyncer(syncer, false, true, zapcore.ErrorLevel)
+		cfg.Logger = logging.NewZapLogger(syncer, false, true, zapcore.ErrorLevel)
 	}
 
-	rr, err := configureRouter(listenerAddr, cfg, &routerConfig, cdn, natsData.Server, zapLogger)
+	if cfg.AccessLogger == nil {
+		cfg.AccessLogger = cfg.Logger
+	}
+
+	rr, err := configureRouter(listenerAddr, cfg, &routerConfig, cdn, natsData.Server)
 	if err != nil {
 		return nil, err
 	}
@@ -564,7 +572,7 @@ func createTestEnv(t testing.TB, cfg *Config) (*Environment, error) {
 	return e, nil
 }
 
-func configureRouter(listenerAddr string, testConfig *Config, routerConfig *nodev1.RouterConfig, cdn *httptest.Server, natsServer *natsserver.Server, zapLogger *zap.Logger) (*core.Router, error) {
+func configureRouter(listenerAddr string, testConfig *Config, routerConfig *nodev1.RouterConfig, cdn *httptest.Server, natsServer *natsserver.Server) (*core.Router, error) {
 	cfg := config.Config{
 		Graph: config.Graph{},
 		CDN: config.CDNConfiguration{
@@ -613,6 +621,7 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 		MaxConcurrentResolvers:         32,
 		ExecutionPlanCacheSize:         1024,
 		EnablePersistedOperationsCache: true,
+		OperationHashCacheSize:         2048,
 		ParseKitPoolSize:               8,
 		EnableValidationCache:          true,
 		ValidationCacheSize:            1024,
@@ -646,7 +655,11 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 	}
 
 	routerOpts := []core.Option{
-		core.WithLogger(zapLogger),
+		core.WithLogger(testConfig.Logger),
+		core.WithAccessLogs(&core.AccessLogsConfig{
+			Logger:     testConfig.AccessLogger,
+			Attributes: testConfig.AccessLogFields,
+		}),
 		core.WithGraphApiToken(graphApiToken),
 		core.WithDevelopmentMode(true),
 		core.WithPlayground(true),
@@ -696,6 +709,7 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 				Exporters:             []config.TracingExporter{},
 				Propagation:           testConfig.PropagationConfig,
 				TracingGlobalFeatures: config.TracingGlobalFeatures{},
+				ResponseTraceHeader:   testConfig.ResponseTraceHeader,
 			},
 		})
 
@@ -783,6 +797,7 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 			testConfig.ModifyWebsocketConfiguration(wsConfig)
 		}
 		routerOpts = append(routerOpts, core.WithWebSocketConfiguration(wsConfig))
+		routerOpts = append(routerOpts, core.WithClientHeader(testConfig.ClientHeader))
 	}
 	return core.NewRouter(routerOpts...)
 }