+++ router/core/graphql_handler.go
@@ -6,22 +6,19 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	rErrors "github.com/wundergraph/cosmo/router/internal/errors"
 	rotel "github.com/wundergraph/cosmo/router/pkg/otel"
+	"go.opentelemetry.io/otel/attribute"
 	"io"
 	"net/http"
 	"strings"
 
-	"go.opentelemetry.io/otel/attribute"
-
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/graphqlerrors"
 
-	"github.com/go-chi/chi/v5/middleware"
 	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
 
 	"github.com/wundergraph/cosmo/router/pkg/config"
-	"github.com/wundergraph/cosmo/router/pkg/logging"
-
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/engine/plan"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/engine/resolve"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/operationreport"
@@ -140,8 +137,7 @@ type GraphQLHandler struct {
 }
 
 func (h *GraphQLHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	requestLogger := h.log.With(logging.WithRequestID(middleware.GetReqID(r.Context())))
-	operationCtx := getOperationContext(r.Context())
+	requestContext := getRequestContext(r.Context())
 
 	var baseAttributes []attribute.KeyValue
 
@@ -156,16 +152,16 @@ func (h *GraphQLHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	defer graphqlExecutionSpan.End()
 
 	ctx := &resolve.Context{
-		Variables: operationCtx.Variables(),
-		Files:     operationCtx.Files(),
+		Variables: requestContext.operation.variables,
+		Files:     requestContext.operation.files,
 		Request: resolve.Request{
 			Header: r.Header,
 		},
 		RenameTypeNames:  h.executor.RenameTypeNames,
-		TracingOptions:   operationCtx.traceOptions,
-		InitialPayload:   operationCtx.initialPayload,
-		Extensions:       operationCtx.extensions,
-		ExecutionOptions: operationCtx.executionOptions,
+		TracingOptions:   requestContext.operation.traceOptions,
+		InitialPayload:   requestContext.operation.initialPayload,
+		Extensions:       requestContext.operation.extensions,
+		ExecutionOptions: requestContext.operation.executionOptions,
 	}
 
 	ctx = ctx.WithContext(executionContext)
@@ -178,60 +174,67 @@ func (h *GraphQLHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 	ctx = h.configureRateLimiting(ctx)
 
-	defer propagateSubgraphErrors(ctx, requestLogger)
-
-	switch p := operationCtx.preparedPlan.preparedPlan.(type) {
+	switch p := requestContext.operation.preparedPlan.preparedPlan.(type) {
 	case *plan.SynchronousResponsePlan:
 		w.Header().Set("Content-Type", "application/json")
-		h.setDebugCacheHeaders(w, operationCtx)
+		h.setDebugCacheHeaders(w, requestContext.operation)
+
 		if h.enableResponseHeaderPropagation {
 			ctx = WithResponseHeaderPropagation(ctx)
 		}
+
+		defer propagateSubgraphErrors(ctx)
+
 		resp, err := h.executor.Resolver.ResolveGraphQLResponse(ctx, p.Response, nil, HeaderPropagationWriter(w, ctx.Context()))
+		requestContext.dataSources = p.Response.DataSources
+
 		if err != nil {
-			requestLogger.Error("unable to resolve response", zap.Error(err))
 			trackResponseError(ctx.Context(), err)
 			h.WriteError(ctx, err, p.Response, w)
 			return
 		}
+
 		graphqlExecutionSpan.SetAttributes(rotel.WgAcquireResolverWaitTimeMs.Int64(resp.ResolveAcquireWaitTime.Milliseconds()))
 	case *plan.SubscriptionResponsePlan:
 		var (
 			writer resolve.SubscriptionResponseWriter
 			ok     bool
 		)
-		h.setDebugCacheHeaders(w, operationCtx)
+		h.setDebugCacheHeaders(w, requestContext.operation)
+
+		defer propagateSubgraphErrors(ctx)
 		ctx, writer, ok = GetSubscriptionResponseWriter(ctx, ctx.Variables, r, w)
 		if !ok {
-			requestLogger.Error("unable to get subscription response writer", zap.Error(errCouldNotFlushResponse))
+			requestContext.logger.Error("unable to get subscription response writer", zap.Error(errCouldNotFlushResponse))
 			trackResponseError(r.Context(), errCouldNotFlushResponse)
-			writeRequestErrors(r, w, http.StatusInternalServerError, graphqlerrors.RequestErrorsFromError(errCouldNotFlushResponse), requestLogger)
+			writeRequestErrors(r, w, http.StatusInternalServerError, graphqlerrors.RequestErrorsFromError(errCouldNotFlushResponse), requestContext.logger)
 			return
 		}
+
 		h.websocketStats.ConnectionsInc()
 		defer h.websocketStats.ConnectionsDec()
 
 		err := h.executor.Resolver.ResolveGraphQLSubscription(ctx, p.Response, writer)
 		if err != nil {
 			if errors.Is(err, context.Canceled) {
-				requestLogger.Debug("context canceled: unable to resolve subscription response", zap.Error(err))
+				requestContext.logger.Debug("context canceled: unable to resolve subscription response", zap.Error(err))
 				trackResponseError(r.Context(), err)
 				return
 			} else if errors.Is(err, ErrUnauthorized) {
 				trackResponseError(ctx.Context(), err)
-				writeRequestErrors(r, w, http.StatusUnauthorized, graphqlerrors.RequestErrorsFromError(err), requestLogger)
+				writeRequestErrors(r, w, http.StatusUnauthorized, graphqlerrors.RequestErrorsFromError(err), requestContext.logger)
 				return
 			}
 
-			requestLogger.Error("unable to resolve subscription response", zap.Error(err))
+			requestContext.logger.Error("unable to resolve subscription response", zap.Error(err))
 			trackResponseError(ctx.Context(), err)
-			writeRequestErrors(r, w, http.StatusInternalServerError, graphqlerrors.RequestErrorsFromError(errCouldNotResolveResponse), requestLogger)
+			writeRequestErrors(r, w, http.StatusInternalServerError, graphqlerrors.RequestErrorsFromError(errCouldNotResolveResponse), requestContext.logger)
 			return
 		}
 	default:
-		requestLogger.Error("unsupported plan kind")
+		requestContext.logger.Error("unsupported plan kind")
 		trackResponseError(ctx.Context(), errOperationPlanUnsupported)
-		writeRequestErrors(r, w, http.StatusInternalServerError, graphqlerrors.RequestErrorsFromError(errOperationPlanUnsupported), requestLogger)
+		writeRequestErrors(r, w, http.StatusInternalServerError, graphqlerrors.RequestErrorsFromError(errOperationPlanUnsupported), requestContext.logger)
 	}
 }
 
@@ -265,12 +268,21 @@ func (h *GraphQLHandler) configureRateLimiting(ctx *resolve.Context) *resolve.Co
 // @TODO This function should be refactored to be a helper function for websocket and http error writing
 // In the websocket case, we call this function concurrently as part of the polling loop. This is error-prone.
 func (h *GraphQLHandler) WriteError(ctx *resolve.Context, err error, res *resolve.GraphQLResponse, w io.Writer) {
-	requestLogger := h.log.With(logging.WithRequestID(middleware.GetReqID(ctx.Context())))
+	reqContext := getRequestContext(ctx.Context())
+
+	if reqContext == nil {
+		h.log.Error("unable to get request context")
+		return
+	}
+
+	requestLogger := reqContext.logger
+
 	httpWriter, isHttpResponseWriter := w.(http.ResponseWriter)
 	response := GraphQLErrorResponse{
 		Errors: make([]graphqlError, 1),
 		Data:   nil,
 	}
+
 	switch getErrorType(err) {
 	case errorTypeRateLimit:
 		response.Errors[0].Message = "Rate limit exceeded"
@@ -348,6 +360,7 @@ func (h *GraphQLHandler) WriteError(ctx *resolve.Context, err error, res *resolv
 			httpWriter.WriteHeader(http.StatusInternalServerError)
 		}
 	}
+
 	if ctx.TracingOptions.Enable && ctx.TracingOptions.IncludeTraceOutputInResponseExtensions {
 		traceNode := resolve.GetTrace(ctx.Context(), res.Fetches)
 		if response.Extensions == nil {
@@ -355,13 +368,19 @@ func (h *GraphQLHandler) WriteError(ctx *resolve.Context, err error, res *resolv
 		}
 		response.Extensions.Trace, err = json.Marshal(traceNode)
 		if err != nil {
-			requestLogger.Error("unable to marshal trace node", zap.Error(err))
+			requestLogger.Error("Unable to marshal trace node", zap.Error(err))
 		}
 	}
+
 	err = json.NewEncoder(w).Encode(response)
 	if err != nil {
-		requestLogger.Error("unable to write rate limit response", zap.Error(err))
+		if rErrors.IsBrokenPipe(err) {
+			requestLogger.Warn("Broken pipe, unable to write error response", zap.Error(err))
+		} else {
+			requestLogger.Error("Unable to write error response", zap.Error(err))
+		}
 	}
+
 	if wsRw, ok := w.(*websocketResponseWriter); ok {
 		_ = wsRw.Flush()
 	}