+++ router/CHANGELOG.md
@@ -4,6 +4,57 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.121.1](https://github.com/wundergraph/cosmo/compare/router@0.121.0...router@0.121.1) (2024-10-07)
+
+**Note:** Version bump only for package router
+
+# [0.121.0](https://github.com/wundergraph/cosmo/compare/router@0.120.0...router@0.121.0) (2024-10-07)
+
+### Features
+
+* allow specified fields in subgraph errors ([#1248](https://github.com/wundergraph/cosmo/issues/1248)) ([72c770d](https://github.com/wundergraph/cosmo/commit/72c770d1d4d795f70586dfaa1a0ffab2943638a4)) (@thisisnithin)
+
+# [0.120.0](https://github.com/wundergraph/cosmo/compare/router@0.119.1...router@0.120.0) (2024-10-07)
+
+### Bug Fixes
+
+* don't set empty cache control headers ([#1246](https://github.com/wundergraph/cosmo/issues/1246)) ([eb1ef06](https://github.com/wundergraph/cosmo/commit/eb1ef068a911eb36a813b9bc718f1a01db62785f)) (@df-wg)
+* improve error logs ([#1247](https://github.com/wundergraph/cosmo/issues/1247)) ([19e5ba6](https://github.com/wundergraph/cosmo/commit/19e5ba62fd1dfe6167b66afc1123477ebdd83381)) (@StarpTech)
+
+### Features
+
+* fallback storage for execution config ([#1241](https://github.com/wundergraph/cosmo/issues/1241)) ([9704342](https://github.com/wundergraph/cosmo/commit/97043429bfadd69a86a560e2105dc8f641d6ad65)) (@thisisnithin)
+
+## [0.119.1](https://github.com/wundergraph/cosmo/compare/router@0.119.0...router@0.119.1) (2024-10-03)
+
+### Bug Fixes
+
+* apollo compatibility for truncating floats, invalid __typenames ([#1242](https://github.com/wundergraph/cosmo/issues/1242)) ([a773ea5](https://github.com/wundergraph/cosmo/commit/a773ea53d7f9abe48033dad822f4424a31e28af2)) (@devsergiy)
+
+# [0.119.0](https://github.com/wundergraph/cosmo/compare/router@0.118.0...router@0.119.0) (2024-10-03)
+
+### Bug Fixes
+
+* fix panic when accessController is nil in handleUpgradeRequest ([#1239](https://github.com/wundergraph/cosmo/issues/1239)) ([5699c03](https://github.com/wundergraph/cosmo/commit/5699c03db4d0b13b39643787700e49a389cf038d)) (@ElliottZeroFlucs)
+
+### Features
+
+* add option to add response header with trace id ([#1234](https://github.com/wundergraph/cosmo/issues/1234)) ([f8f5078](https://github.com/wundergraph/cosmo/commit/f8f50781917ee4bb0c24ac0c8aa9db71f87e9a05)) (@JivusAyrus)
+* add option to have custom names client name and version headers ([#1233](https://github.com/wundergraph/cosmo/issues/1233)) ([e09348d](https://github.com/wundergraph/cosmo/commit/e09348d2925469e259b526c231a14c6abcc80916)) (@JivusAyrus)
+* router version command ([#1240](https://github.com/wundergraph/cosmo/issues/1240)) ([39970e5](https://github.com/wundergraph/cosmo/commit/39970e5b1c6d8e2a2958f5ea55bb0d8aa7206d92)) (@StarpTech)
+
+# [0.118.0](https://github.com/wundergraph/cosmo/compare/router@0.117.0...router@0.118.0) (2024-10-03)
+
+### Features
+
+* advanced access logs ([#1203](https://github.com/wundergraph/cosmo/issues/1203)) ([b4dc9ac](https://github.com/wundergraph/cosmo/commit/b4dc9acd964b2982fa9e27ec066e91583991ee17)) (@StarpTech)
+
+# [0.117.0](https://github.com/wundergraph/cosmo/compare/router@0.116.3...router@0.117.0) (2024-10-02)
+
+### Features
+
+* add apollo compatibility flag to truncate floats ([#1236](https://github.com/wundergraph/cosmo/issues/1236)) ([17c80c9](https://github.com/wundergraph/cosmo/commit/17c80c91f4e594cf48dae887840517b588bbe0f8)) (@jensneuse)
+
 ## [0.116.3](https://github.com/wundergraph/cosmo/compare/router@0.116.2...router@0.116.3) (2024-09-30)
 
 ### Bug Fixes